# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Test tracking and run management REST API.",
        "title": "TTGO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/run-folders": {
            "get": {
                "description": "Return all run folders as a flat list (backward compat) or tree (?view=tree)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "List run folders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set to 'tree' for hierarchical response",
                        "name": "view",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "run_folders array",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new run folder with the given name, optionally nested under a parent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "Create a run folder",
                "parameters": [
                    {
                        "description": "Run folder name and optional parent_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "parent_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ttgo_pkg_tracker_models.RunFolder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/run-folders/{id}": {
            "delete": {
                "description": "Delete an existing run folder and all its subfolders by ID",
                "tags": [
                    "run-folders"
                ],
                "summary": "Delete a run folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update the name of an existing run folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "Rename a run folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/run-folders/{id}/copy": {
            "post": {
                "description": "Deep-copy a run folder with all subfolders and runs (results reset to PENDING)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "Copy a run folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional new name and parent",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "parent_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ttgo_pkg_tracker_models.RunFolder"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/run-folders/{id}/order": {
            "patch": {
                "description": "Update the display order of an existing run folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "Reorder a run folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New display order",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "display_order": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/run-folders/{id}/parent": {
            "patch": {
                "description": "Move a run folder to a new parent (or to root if parent_id is null)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "run-folders"
                ],
                "summary": "Move a run folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run folder ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New parent_id (null for root)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "parent_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ttgo_pkg_tracker_models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "PASS",
                "FAIL",
                "SKIP",
                "PENDING",
                "RUNNING",
                "ERROR"
            ],
            "x-enum-varnames": [
                "StatusPass",
                "StatusFail",
                "StatusSkip",
                "StatusPending",
                "StatusRunning",
                "StatusError"
            ]
        },
...
        "ttgo_pkg_tracker_models.RunFolder": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "Nullable for root folders",
                    "type": "string"
                },
                "sub_folders": {
                    "description": "Children fields for API responses (not stored in DB directly)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ttgo_pkg_tracker_models.RunFolder"
                    }
                },
                "test_runs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ttgo_pkg_tracker_models.TestRun"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "ttgo_pkg_tracker_models.RunResult": {
            "type": "object",
            "properties": {
                "app_version": {
                    "type": "string"
                },
                "attempt_number": {
                    "type": "integer"
                },
                "browser": {
                    "description": "Context",
                    "type": "string"
                },
                "closed_defect_link_count": {
                    "type": "integer"
                },
                "created_at": {
                    "description": "Timestamps",
                    "type": "string"
                },
                "defect_type": {
                    "description": "DefectType classifies the failure reason for FAIL results.\nValues: \"to_investigate\" | \"product_bug\" | \"automation_bug\" | \"system_issue\" | \"\" (not applicable)",
                    "type": "string"
                },
                "duration_ms": {
                    "description": "Timing",
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "environment": {
                    "type": "string"
                },
                "error_message": {
                    "description": "Failure Details",
                    "type": "string"
                },
                "failure_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "log_text": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "open_defect_link_count": {
                    "description": "Defect link counts (not persisted, populated by GetTestRun)",
                    "type": "integer"
                },
                "os": {
                    "type": "string"
                },
                "screenshots": {
                    "description": "Artifacts",
                    "type": "string"
                },
                "stack_trace": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/ttgo_pkg_tracker_models.ExecutionStatus"
                },
                "steps": {
                    "description": "Detailed Steps \u0026 Metadata",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "test_case": {
                    "description": "Association (not stored, loaded via preload)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ttgo_pkg_tracker_models.TestCase"
                        }
                    ]
                },
                "test_case_id": {
                    "description": "Nullable if test case is deleted",
                    "type": "string"
                },
                "test_name_snapshot": {
                    "description": "Preserved name",
                    "type": "string"
                },
                "test_run_id": {
                    "type": "string"
                },
                "trace_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "ttgo_pkg_tracker_models.TestRun": {
            "type": "object",
            "properties": {
                "automation_bug_count": {
                    "type": "integer"
                },
                "category_id": {
                    "description": "nullable — empty runs have no category",
                    "type": "string"
                },
                "closed_defect_link_count": {
                    "type": "integer"
                },
                "comment_count": {
                    "description": "Comment count (not persisted, populated by list handler)",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "failed_results": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "open_defect_link_count": {
                    "description": "Defect link counts (not persisted, populated by list handler)",
                    "type": "integer"
                },
                "passed_results": {
                    "type": "integer"
                },
                "pending_results": {
                    "type": "integer"
                },
                "product_bug_count": {
                    "type": "integer"
                },
                "retried_count": {
                    "description": "Retry stats (not persisted, populated by GetTestRun/GetTestRuns)",
                    "type": "integer"
                },
                "run_folder_id": {
                    "description": "nullable FK to RunFolder",
                    "type": "string"
                },
                "run_results": {
                    "description": "Associations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ttgo_pkg_tracker_models.RunResult"
                    }
                },
                "skipped_results": {
                    "type": "integer"
                },
                "status": {
                    "description": "PENDING, RUNNING, PASSED, FAILED, etc.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ttgo_pkg_tracker_models.ExecutionStatus"
                        }
                    ]
                },
                "system_issue_count": {
                    "type": "integer"
                },
                "to_investigate_count": {
                    "description": "Defect type breakdowns (not persisted, populated by GetTestRuns list query)",
                    "type": "integer"
                },
                "total_attempts": {
                    "type": "integer"
                },
                "total_results": {
                    "description": "Computed aggregate fields (not persisted, populated by GetTestRuns list query)",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: \"Bearer \u003cyour-token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}